requirements.txt
PySide6==6.3.0
PyYAML==6.0
bandit==1.7.4
pygount==1.4.0
radon==5.1.0
rich==12.4.4
